Molts canvis.

+++ src/Engine/Core/ISubSystem.h
@@ -1,5 +1,6 @@
 #ifndef ISUBSYSTEM_H_
 #define ISUBSYSTEM_H_
+#include "CBasicTypes.h"
 
 class ISubSystem
 {
@@ -8,7 +9,7 @@ public:
 
 	virtual void StartUp(void) = 0;
 	virtual void ShutDown(void) = 0;
-	virtual void Update(float dt) = 0;
+	virtual void Update(f64 dt) = 0;
 };
 
 #endif
